requirements.txt
python-dotenv
langchain
pinecone-client
PyPDF2
gradio
openai
tiktoken
